+++ src/Commands/ConsoleCmdAmnesia.cs
@@ -23,6 +23,7 @@ namespace Amnesia.Commands
                 { "players", "show players and their amnesia-related info" },
                 { "grant <user id / player name / entity id> <timeInSeconds>", "grant player some bonus xp time" },
                 { "fragile <user id / player name / entity id> <true/false>", "give or remove fragile memory debuff" },
+                { "records <user id / player name / entity id>", "show skill/perk records in order they were purchased by the given player" },
                 { "config", "show current amnesia configuration" },
                 { "set", "show the single-value fields you can adjust" },
                 { "set <field>", "describe how you can update this field" },
@@ -89,6 +90,9 @@ namespace Amnesia.Commands
                             break;
                         }
                         HandleFragile(_params);
+                        return;
+                    case "records":
+                        HandleRecords(_params);
                         return;
                     case "config":
                         SdtdConsole.Instance.Output(Config.AsString());
@@ -191,6 +195,30 @@ namespace Amnesia.Commands
             }
             player.Buffs.RemoveBuff(Values.BuffFragileMemory);
             SdtdConsole.Instance.Output($"Successfully removed {Values.BuffFragileMemory} from {player.GetDebugName()}.");
+        }
+
+        private void HandleRecords(List<string> @params)
+        {
+            var clientInfo = ConsoleHelper.ParseParamIdOrName(@params[1], true, false);
+            if (clientInfo == null || !GameManager.Instance.World.Players.dict.TryGetValue(clientInfo.entityId, out var player))
+            {
+                SdtdConsole.Instance.Output("Unable to find this player; note: player must be online");
+                return;
+            }
+            if (!PlayerRecord.Entries.TryGetValue(clientInfo.entityId, out var playerRecord))
+            {
+                SdtdConsole.Instance.Output($"Unable to find an active record for player {clientInfo.entityId}");
+                return;
+            }
+            if (playerRecord.Changes.Count == 0)
+            {
+                SdtdConsole.Instance.Output($"Player {clientInfo.entityId} had no recorded changes.");
+                return;
+            }
+            for (int i = 0; i < playerRecord.Changes.Count; i++)
+            {
+                SdtdConsole.Instance.Output($"{i+1,3}. {playerRecord.Changes[i].Item1} >> {playerRecord.Changes[i].Item2}");
+            }
         }
 
         private void RouteListRequest(List<string> @params)