+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+- add console command to view skill tracker
 - add dialog purchasing system for treatments
 - add player skill/perk tracking
 - add success/unnecessary/unaffordable responses